requirements.txt
Django==4.1
django-jazzmin
django-bootstrap5
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
django-rest-swagger